mysql: fix dispose-method not called in external-mode
