--- pkg/tsdb/mysql/standalone/datasource.go
@@ -1,32 +1,0 @@
-package main
-
-import (
-	"context"
-
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
-	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
-	mysqlds "github.com/grafana/grafana/pkg/tsdb/mysql"
-)
-
-var (
-	_ backend.QueryDataHandler   = (*Datasource)(nil)
-	_ backend.CheckHealthHandler = (*Datasource)(nil)
-)
-
-func NewDatasource(context.Context, backend.DataSourceInstanceSettings) (instancemgmt.Instance, error) {
-	return &Datasource{
-		Service: mysqlds.ProvideService(),
-	}, nil
-}
-
-type Datasource struct {
-	Service *mysqlds.Service
-}
-
-func (d *Datasource) QueryData(ctx context.Context, req *backend.QueryDataRequest) (*backend.QueryDataResponse, error) {
-	return d.Service.QueryData(ctx, req)
-}
-
-func (d *Datasource) CheckHealth(ctx context.Context, req *backend.CheckHealthRequest) (*backend.CheckHealthResult, error) {
-	return d.Service.CheckHealth(ctx, req)
-}
